Extract logic from keymaps into sibling actions modules
Keymap files now contain only key bindings; all non-trivial functions
live in adjacent actions.lua modules. The new convention:

  <dir>/keymaps.lua  -- pure vim.keymap.set calls referencing actions
  <dir>/actions.lua  -- function definitions imported by the keymaps

Applied to root config, lsp/, and plugins that previously embedded
logic inline (telescope, orgmode, gitlinker, neotest). Plugin keymap
files that were already pure command strings (diffview, fugitive,
oil, etc.) are unchanged.
