+++ config/nvim/init.lua
@@ -14,6 +14,4 @@ require("lsp")
 require("keymaps")
 
 -- Create user command for lazydocker
-vim.api.nvim_create_user_command("Lazydocker", function()
-	require("tools.lazydocker").toggle_lazydocker()
-end, {})
+vim.api.nvim_create_user_command("Lazydocker", require("actions").toggle_lazydocker, {})
