+++ config/nvim/plugins/gitlinker/keymaps.lua
@@ -1,9 +1,8 @@
 local M = {}
 
 function M.setup()
-	vim.keymap.set("n", "<leader>gy", function()
-		require("gitlinker").get_buf_range_url("n")
-	end, { desc = "Copy git link to clipboard" })
+	local actions = require("plugins.gitlinker.actions")
+	vim.keymap.set("n", "<leader>gy", actions.copy_git_link, { desc = "Copy git link to clipboard" })
 end
 
 return M
